Create ByteArray section and popluate it.
Also, make more explicit where "check_requires" comes from.
It was imported the wrong way, and we don't do this kind of things
here. Implicit imports make sense only at the top level.
