+++ test/builtin/numbers/test_calculus.py
@@ -3,10 +3,9 @@
 FindRoot[], FindMinimum[], NFindMaximum[] tests
 
 """
-import importlib
 import pytest
 from test.helper import evaluate
-from mathics.builtin import check_requires_list
+from mathics.builtin.base import check_requires_list
 
 
 if check_requires_list(["scipy", "scipy.integrate"]):
